+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- Add a `.no-article-found` style to center the message.
 - Update the `parts/tag-header.php` definition appearance
 - Update `template/tempalte-query-loop` to show "Articoli Correlati" as a title for "tags" query loops.
 - Change `.tag-readings` font style to italics
 ### Fixed
-- Fix `.author-bio-teaser .more-link-toggle` margin
+- Fix author's bio menu toggle appearance
 - Fix `.tag-title` appearance
 - Fix `.post-tags` style and link clickability