+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.1] 2025-08-17
 ### Changed
 - Update tag readings appearance to collapsable pills.
 - Add a `.no-article-found` style to center the message.
@@ -119,7 +119,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.7.1...HEAD
+[0.7.1]: https://github.com/erremauro/cigno-zen/releases/tag/v0.7.1
 [0.7.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.7.0
 [0.6.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.6.0
 [0.5.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.5.0